pyproject.toml
[tool.poetry]
name = "fumocam-discordbot"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.19.2"
pytz = "^2021.3"
discord-py = "^2.3.2"
aiohttp = ">=3.8.5"       # CVE Fix

[tool.poetry.group.dev.dependencies]
black = "^22.1.0"
bandit = "^1.7.4"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.931"
ossaudit = "^0.5.0"
types-pytz = "^2021.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Customizations
[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
show_error_codes = true
